+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Official Docker images released - [see here](https://hub.docker.com/r/quentinblampey/novae)
 - Support `reference=None` in `fine_tune` (random initialization) and used as default
 - Added `fast` argument in `novae.plot.domains` to have a quick (but less accurate) rendering of domains.
+- Added `novae.settings.scale_to_microns` if the coordinates are not in microns
 
 ### Breaking changes
 - Remove support for `python==3.10`